requirements.txt
# ── Core ML & Data ────────────────────────────────────────────────────────────
torch>=2.3.0
numpy>=1.26.0
pandas>=2.2.0
scikit-learn>=1.5.0
scipy>=1.13.0

# ── HuggingFace ecosystem ──────────────────────────────────────────────────────
transformers>=4.44.0
datasets>=2.20.0
huggingface-hub>=0.24.0
sentence-transformers>=3.0.0
accelerate>=0.33.0
peft>=0.12.0

# ── Retrieval & Search ─────────────────────────────────────────────────────────
rank-bm25>=0.2.2
faiss-cpu>=1.8.0

# ── LLM / Generation ──────────────────────────────────────────────────────────
openai>=1.40.0
anthropic>=0.34.0
tiktoken>=0.7.0

# ── Evaluation ────────────────────────────────────────────────────────────────
nltk>=3.8.1
rouge-score>=0.1.2
sacrebleu>=2.4.2

# ── Experiment Tracking ───────────────────────────────────────────────────────
wandb>=0.17.0
mlflow>=2.15.0

# ── Notebooks & Visualisation ─────────────────────────────────────────────────
jupyterlab>=4.2.0
ipywidgets>=8.1.0
matplotlib>=3.9.0
seaborn>=0.13.0
plotly>=5.23.0

# ── Utilities ─────────────────────────────────────────────────────────────────
tqdm>=4.66.0
python-dotenv>=1.0.0
pyyaml>=6.0.1
rich>=13.7.0
loguru>=0.7.2
typer>=0.12.0
hydra-core>=1.3.2
omegaconf>=2.3.0
